CORE-7637: moved the unit and integration tests for the permissions service to a separate namespace
