+++ golang/src/permissions/restapi/impl/test/db_test.go
@@ -1,4 +1,4 @@
-package impl
+package test
 
 import (
 	"database/sql"
